VMS-21173: Add support of std::variant in Server API

+++ vms/libs/nx_vms_json_rpc/src/nx/vms/json_rpc/detail/outgoing_processor.h
@@ -20,7 +20,7 @@ namespace nx::vms::json_rpc::detail {
 class OutgoingProcessor
 {
 public:
-    using Id = std::variant<int64_t, QString>;
+    using Id = std::variant<int, QString>;
     using SendFunc = nx::utils::MoveOnlyFunc<void(QJsonValue)>;
     OutgoingProcessor(SendFunc sendFunc): m_sendFunc(std::move(sendFunc)) {}
